requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
SQLAlchemy==2.0.48
requests==2.32.5
beautifulsoup4==4.14.3
googlesearch-python==1.3.0
python-multipart==0.0.22
APScheduler==3.11.2
